+++ codex/serializers/browser/filters.py
@@ -1,14 +1,18 @@
 """Browser Settings Filter Serializers."""
 
-from rest_framework.serializers import ListField, Serializer
+from rest_framework.serializers import Serializer
 
 from codex.serializers.fields import (
     VuetifyBooleanField,
     VuetifyDecadeField,
     VuetifyDecimalField,
-    VuetifyIntegerField,
 )
 from codex.serializers.fields.browser import BookmarkFilterField
+from codex.serializers.fields.vuetify import (
+    VuetifyFileTypeChoiceField,
+    VuetifyListField,
+    VuetifyReadingDirectionChoiceField,
+)
 
 
 class BrowserSettingsFilterInputSerializer(Serializer):
@@ -16,27 +20,27 @@ class BrowserSettingsFilterInputSerializer(Serializer):
 
     bookmark = BookmarkFilterField(required=False)
     # Dynamic filters
-    age_rating = ListField(child=VuetifyIntegerField(), required=False)
-    characters = ListField(child=VuetifyIntegerField(), required=False)
-    country = ListField(child=VuetifyIntegerField(), required=False)
-    credits = ListField(child=VuetifyIntegerField(), required=False)
-    critical_rating = ListField(
-        child=VuetifyDecimalField(max_digits=5, decimal_places=2), required=False
+    age_rating = VuetifyListField()
+    characters = VuetifyListField()
+    country = VuetifyListField()
+    credits = VuetifyListField()
+    critical_rating = VuetifyListField(
+        child=VuetifyDecimalField(max_digits=5, decimal_places=2)
     )
-    decade = ListField(child=VuetifyDecadeField(), required=False)
-    file_type = ListField(child=VuetifyIntegerField(), required=False)
-    genres = ListField(child=VuetifyIntegerField(), required=False)
-    identifier_source = ListField(child=VuetifyIntegerField(), required=False)
-    language = ListField(child=VuetifyIntegerField(), required=False)
-    locations = ListField(child=VuetifyIntegerField(), required=False)
-    monochrome = ListField(child=VuetifyBooleanField(), required=False)
-    original_format = ListField(child=VuetifyIntegerField(), required=False)
-    reading_direction = ListField(child=VuetifyIntegerField(), required=False)
-    series_groups = ListField(child=VuetifyIntegerField(), required=False)
-    stories = ListField(child=VuetifyIntegerField(), required=False)
-    story_arcs = ListField(child=VuetifyIntegerField(), required=False)
-    tagger = ListField(child=VuetifyIntegerField(), required=False)
-    tags = ListField(child=VuetifyIntegerField(), required=False)
-    teams = ListField(child=VuetifyIntegerField(), required=False)
-    universes = ListField(child=VuetifyIntegerField(), required=False)
-    year = ListField(child=VuetifyIntegerField(), required=False)
+    decade = VuetifyListField(child=VuetifyDecadeField)
+    file_type = VuetifyListField(child=VuetifyFileTypeChoiceField)
+    genres = VuetifyListField()
+    identifier_source = VuetifyListField()
+    language = VuetifyListField()
+    locations = VuetifyListField()
+    monochrome = VuetifyListField(child=VuetifyBooleanField)
+    original_format = VuetifyListField()
+    reading_direction = VuetifyListField(child=VuetifyReadingDirectionChoiceField)
+    series_groups = VuetifyListField()
+    stories = VuetifyListField()
+    story_arcs = VuetifyListField()
+    tagger = VuetifyListField()
+    tags = VuetifyListField()
+    teams = VuetifyListField()
+    universes = VuetifyListField()
+    year = VuetifyListField()